document a few more cargo features
